backend/pyproject.toml
[tool.poetry]
name = "task_crud_backend"
version = "0.1.0"
description = "Backend for task CRUD operations with authentication"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.0"
uvicorn = "^0.32.0"
sqlmodel = "^0.0.22"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-multipart = "^0.0.20"
bcrypt = "^4.2.0"
pydantic-settings = "^2.0.0"
psycopg2-binary = "^2.9.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^5.0.0"
black = "^24.0.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"
isort = "^5.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "src.main:app"

[[tool.poetry.packages]]
include = "src"

[tool.black]
line-length = 88
target-version = ['py313']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.flake8]
max-line-length = 88
extend-ignore = [
    "E203",  # whitespace before ':'
    "W503",  # line break before binary operator
]